change pytorch integer dtype to 32

+++ muograph/utils/params.py
@@ -4,6 +4,6 @@ dtype_hit = torch.float32
 
 dtype_track = torch.float32
 
-dtype_n = torch.int16
+dtype_n = torch.int32
 
 dtype_E = torch.float64
